osarebito-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
email-validator
SQLAlchemy
psycopg2-binary
Pillow
